all_ok

+++ Libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: floris <floris@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/12 21:26:16 by fleonte           #+#    #+#             */
-/*   Updated: 2023/09/17 01:22:13 by floris           ###   ########.fr       */
+/*   Updated: 2023/10/05 14:58:58 by floris           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (*s++)
